backend/requirements.txt
Flask>=2.0
Flask-Cors>=3.0.10
Flask-SocketIO>=5.3.6
torch>=1.9.0
numpy>=1.19.0
supabase>=2.0.0
python-dotenv>=1.0.0